requirements.txt
bcrypt==5.0.0
blinker==1.9.0
click==8.3.1
dlib==20.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.1.2
flask-cors==6.0.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.1
pillow==12.1.0
PyJWT==2.10.1
setuptools==80.9.0
Werkzeug==3.1.5
wheel==0.45.1
